fix add blog page
