+++ Web/starter_project_iconics/pages/blog/add-blog.tsx
@@ -1,48 +1,9 @@
+import TextEditor from '@/components/blog/TextEditor'
 import Image from 'next/image'
-import React, { useState } from 'react'
-import ReactQuill from 'react-quill'
+import React from 'react'
 import 'react-quill/dist/quill.snow.css' // import the stylesheet
 
-interface CustomIcons {
-  [key: string]: string
-}
-
 const AddBlog: React.FC = () => {
-  const [content, setContent] = useState('')
-
-  const handleChange = (value: string) => {
-    setContent(value)
-  }
-
-  const editorModules = {
-    toolbar: [
-      ['bold', 'italic', 'underline'],
-      [{ list: 'ordered' }, { list: 'bullet' }],
-      ['link'],
-      ['video', 'image', 'file'],
-    ],
-  }
-
-  const customIcons: CustomIcons = {
-    link: '<i class="fas fa-link"></i>',
-    image: '<i class="fas fa-image"></i>',
-    video: '<i class="fas fa-video"></i>',
-    file: '<i class="fas fa-file"></i>',
-  }
-
-  const toolbarOptions = editorModules.toolbar.map((group) =>
-    group.map((button: any) => {
-      if (typeof button !== 'object' || !('value' in button)) {
-        return button
-      }
-      const customIcon = customIcons[button.value]
-      return {
-        ...button,
-        icon: customIcon !== undefined ? customIcon : button.icon,
-      }
-    })
-  )
-
   return (
     <div className="grid grid-cols-4 gap-8 bg-white">
       <div className="mx-auto w-5/6 col-span-3">
@@ -77,12 +38,7 @@ const AddBlog: React.FC = () => {
           </div>
         </div>
         <div className="border-none">
-          <ReactQuill
-            value={content}
-            onChange={handleChange}
-            modules={{ toolbar: toolbarOptions }}
-            className="mt-12 mb-8 h-48 pb-12"
-          />
+          <TextEditor />
         </div>
         <div className="flex justify-end gap-x-7 font-montserrat font-semibold mb-8 mt-8">
           <a href="#" className="pt-3 text-blue-600">
@@ -94,7 +50,12 @@ const AddBlog: React.FC = () => {
         </div>
       </div>
 
-      <div className="col-span-1 border-l-2 border-indigo-100 h-3/5 mt-28"></div>
+      <div
+        className="col-span-1 border-l-2 border-indigo-100 mt-8"
+        style={{ height: '575px' }}
+      >
+        <h3 className='font-bold ml-4'>Select Tag</h3>
+      </div>
     </div>
   )
 }
